requirements.txt
flask==2.3.3
pymongo==4.5.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
dnspython==2.4.2